requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# LLM
anthropic>=0.40.0

# Frontend
streamlit==1.33.0
requests==2.32.3

# Data processing
pandas==2.2.2
pyarrow>=14.0.0
numpy==1.26.4
scikit-learn==1.4.2
Pillow==10.3.0

# Image embeddings (CLIP) — CPU build; no CUDA required
torch==2.3.0
transformers==4.40.0

# Environment / config
python-dotenv==1.0.1

# Testing
pytest==8.2.0
